backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
bcrypt==4.3.0
python-jose[cryptography]
python-multipart